大規模リファクタリング
リファクタリング詳細
設定を変更してオートフォーマットを全ファイルに実行
空白行の削除など
タイプミス、IntelliJによる警告の解消
フィールド名の修正
NBTのkey名変更
super();の削除
不必要なintをshortやbyteに変更
言語指定名の変更
テクスチャ名の変更
OfalenSmeltingRecipeの変更
コメントのリフォーマット
とき→時
弓の痕跡の削除
G1表記の追加
バランス調整の準備
バージョンチェックの修正
TeleportMarker→TeleportingMarker

+++ main/java/nahama/ofalenmod/entity/EntityExplosionBall.java
@@ -7,18 +7,17 @@ import net.minecraft.util.MovingObjectPosition;
 import net.minecraft.world.World;
 
 public class EntityExplosionBall extends EntityThrowable {
-
-	private int power;
+	private short power;
 
 	public EntityExplosionBall(World world, EntityLivingBase entity, int grade) {
 		super(world, entity);
 		// Configで設定された値のgrade*2倍を爆発力に代入。
-		power = OfalenModConfigCore.sizeExplosion * (grade * 2);
+		power = (short) (OfalenModConfigCore.sizeExplosion * (grade * 2));
 		if (power == 0)
 			power = OfalenModConfigCore.sizeExplosion;
 	}
 
-	/** ブロックかエンティティに当たった時の処理 */
+	/** ブロックかエンティティに当たった時の処理。 */
 	@Override
 	protected void onImpact(MovingObjectPosition position) {
 		if (!worldObj.isRemote) {
@@ -31,5 +30,4 @@ public class EntityExplosionBall extends EntityThrowable {
 			this.setDead();
 		}
 	}
-
 }
